backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
authlib==1.3.0
python-jose==3.3.0
pyjwt==2.8.0
